requirements.txt
cloudify-common==4.5.5
requests==2.21.0
cloudify-utilities-plugins-sdk==0.0.15
